+++ src/main/java/com/cburch/logisim/tools/key/NumericConfigurator.java
@@ -36,18 +36,18 @@ public abstract class NumericConfigurator<V> implements KeyConfigurator, Cloneab
   private static final int MAX_TIME_KEY_LASTS = 800;
 
   private Attribute<V> attr;
-  private int minValue;
-  private int maxValue;
-  private int curValue;
+  private long minValue;
+  private long maxValue;
+  private long curValue;
   private int radix;
   private int modsEx;
   private long whenTyped;
 
-  public NumericConfigurator(Attribute<V> attr, int min, int max, int modifiersEx) {
+  public NumericConfigurator(Attribute<V> attr, long min, long max, int modifiersEx) {
     this(attr, min, max, modifiersEx, 10);
   }
 
-  public NumericConfigurator(Attribute<V> attr, int min, int max, int modifiersEx, int radix) {
+  public NumericConfigurator(Attribute<V> attr, long min, long max, int modifiersEx, int radix) {
     this.attr = attr;
     this.minValue = min;
     this.maxValue = max;
@@ -71,13 +71,13 @@ public abstract class NumericConfigurator<V> implements KeyConfigurator, Cloneab
     }
   }
 
-  protected abstract V createValue(int value);
+  protected abstract V createValue(long value);
 
-  protected int getMaximumValue(AttributeSet attrs) {
+  protected long getMaximumValue(AttributeSet attrs) {
     return maxValue;
   }
 
-  protected int getMinimumValue(AttributeSet attrs) {
+  protected long getMinimumValue(AttributeSet attrs) {
     return minValue;
   }
 
@@ -89,9 +89,9 @@ public abstract class NumericConfigurator<V> implements KeyConfigurator, Cloneab
         long now = System.currentTimeMillis();
         long sinceLast = now - whenTyped;
         AttributeSet attrs = event.getAttributeSet();
-        int min = getMinimumValue(attrs);
-        int max = getMaximumValue(attrs);
-        int val = 0;
+        long min = getMinimumValue(attrs);
+        long max = getMaximumValue(attrs);
+        long val = 0;
         if (sinceLast < MAX_TIME_KEY_LASTS) {
           val = radix * curValue;
           if (val > max) {
